test: share the value types with restricted operations through test/support/utils.h
minmax_element.pass.cpp and value_storable.pass.cpp defined nearly the same
set of value types with one operation each restricted - eight near-duplicate
pairs, about 200 lines. The two sets have to stay in lock-step: a type the
trait accepts has to be one the algorithms actually compile for, which is the
whole point of checking the same types in both tests.

Move them to test/support/utils.h, next to NoDefaultCtorWrapper, and keep the
"*Compare" shape, which is a superset of the other one - it adds a constructor
from std::int32_t and operator<. Regular and OnlyLessCompare were the same
type and are merged; ExplicitCopyCtorCompare is the traits test's local type
under the shared name.

No change in what either test checks. This also removes a latent clash
between the traits test's local MoveOnly and the global one in
test/support/move_only.h.
