+++ test/general/implementation_details/value_storable.pass.cpp
@@ -7,14 +7,14 @@
 //
 //===------------------------------------------------------===//
 
-// Compile-time checks for oneapi::dpl::__unseq_backend::__is_value_storable_v.
+// Compile-time checks for oneapi::dpl::__unseq_backend::__is_value_storable_v, the condition that selects the vector
+// code path of min_element and minmax_element.
 
 #include "support/test_config.h"
 
 #include <oneapi/dpl/pstl/unseq_backend_simd.h>
 
 #include <cstddef>
-#include <initializer_list>
 #include <iterator>
 #include <type_traits>
 #include <utility>
@@ -24,104 +24,10 @@
 
 namespace dpl_unseq = oneapi::dpl::__unseq_backend;
 
-//----------------------------------------------------------------------------//
-// Value types
-//----------------------------------------------------------------------------//
-
-// Satisfies every requirement: default-constructible, copy-constructible, copy-assignable.
-struct Regular
-{
-    int val = 0;
-};
-
-struct ExplicitDefaultCtor
-{
-    int val;
-    explicit ExplicitDefaultCtor() : val(0) {}
-};
-
-// Default-constructible: the aggregate default-initializes its member through its explicit default constructor.
-struct AggregateOfExplicitDefaultCtor
-{
-    ExplicitDefaultCtor member;
-};
-
-// Not default-constructible: its only constructor takes an initializer list.
-struct BraceInitOnly
-{
-    int val;
-    BraceInitOnly(std::initializer_list<int> init) : val(init.size() == 0 ? 0 : *init.begin()) {}
-};
-
-// A type that is not default-constructible is taken from the test utilities: TestUtils::NoDefaultCtorWrapper<int>.
-
-struct NoCopyAssign
-{
-    int val = 0;
-    NoCopyAssign() = default;
-    NoCopyAssign(const NoCopyAssign&) = default;
-    NoCopyAssign&
-    operator=(const NoCopyAssign&) = delete;
-};
-
-// The copy assignment returns void instead of VoidAssign&.
-struct VoidAssign
-{
-    int val = 0;
-    void
-    operator=(const VoidAssign& other)
-    {
-        val = other.val;
-    }
-};
-
-struct MoveOnly
-{
-    int val = 0;
-    MoveOnly() = default;
-    MoveOnly(MoveOnly&&) = default;
-    MoveOnly&
-    operator=(MoveOnly&&) = default;
-    MoveOnly(const MoveOnly&) = delete;
-    MoveOnly&
-    operator=(const MoveOnly&) = delete;
-};
-
-// Copyable, but with deleted move operations.
-struct CopyOnlyNoMove
-{
-    int val = 0;
-    CopyOnlyNoMove() = default;
-    CopyOnlyNoMove(const CopyOnlyNoMove&) = default;
-    CopyOnlyNoMove&
-    operator=(const CopyOnlyNoMove&) = default;
-    CopyOnlyNoMove(CopyOnlyNoMove&&) = delete;
-    CopyOnlyNoMove&
-    operator=(CopyOnlyNoMove&&) = delete;
-};
-
-// Copyable and assignable from a const lvalue only.
-struct ConstCopyOnly
-{
-    int val = 0;
-    ConstCopyOnly() = default;
-    ConstCopyOnly(const ConstCopyOnly&) = default;
-    ConstCopyOnly&
-    operator=(const ConstCopyOnly&) = default;
-    ConstCopyOnly(ConstCopyOnly&) = delete;
-    ConstCopyOnly&
-    operator=(ConstCopyOnly&) = delete;
-};
-
-// The copy constructor is explicit, so the type is copy-constructible, but its elements cannot be copy-initialized.
-struct ExplicitCopyCtor
-{
-    int val = 0;
-    ExplicitCopyCtor() = default;
-    explicit ExplicitCopyCtor(const ExplicitCopyCtor& other) : val(other.val) {}
-    ExplicitCopyCtor&
-    operator=(const ExplicitCopyCtor&) = default;
-};
+// The value types with restricted operations are defined in test/support/utils.h, so that
+// test/parallel_api/algorithm/alg.sorting/alg.min.max/minmax_element.pass.cpp runs the algorithms on the very same set
+// that is checked against the trait here. TestUtils::NoDefaultCtorWrapper<int> is the type that is not
+// default-constructible.
 
 //----------------------------------------------------------------------------//
 // Reference types
@@ -154,30 +60,31 @@ static_assert(dpl_unseq::__is_value_storable_v<int*>);
 static_assert(dpl_unseq::__is_value_storable_v<const int*>);
 static_assert(dpl_unseq::__is_value_storable_v<std::vector<int>::iterator>);
 static_assert(dpl_unseq::__is_value_storable_v<std::vector<int>::const_iterator>);
-static_assert(dpl_unseq::__is_value_storable_v<Regular*>);
-static_assert(dpl_unseq::__is_value_storable_v<ExplicitDefaultCtor*>);
-static_assert(dpl_unseq::__is_value_storable_v<AggregateOfExplicitDefaultCtor*>);
+static_assert(dpl_unseq::__is_value_storable_v<TestUtils::OnlyLessCompare*>);
+static_assert(dpl_unseq::__is_value_storable_v<TestUtils::ExplicitDefaultCtorCompare*>);
+static_assert(dpl_unseq::__is_value_storable_v<TestUtils::AggregateOfExplicitDefaultCtorCompare*>);
 // The requirements are default construction, copy construction and copy assignment, and nothing else.
-static_assert(dpl_unseq::__is_value_storable_v<CopyOnlyNoMove*>);
-static_assert(dpl_unseq::__is_value_storable_v<VoidAssign*>);
-static_assert(dpl_unseq::__is_value_storable_v<const ConstCopyOnly*>);
+static_assert(dpl_unseq::__is_value_storable_v<TestUtils::CopyOnlyNoMoveCompare*>);
+static_assert(dpl_unseq::__is_value_storable_v<TestUtils::VoidAssignCompare*>);
+static_assert(dpl_unseq::__is_value_storable_v<const TestUtils::ConstCopyOnlyCompare*>);
 // The reference type is not part of the requirement.
 static_assert(dpl_unseq::__is_value_storable_v<std::vector<bool>::iterator>);
 static_assert(dpl_unseq::__is_value_storable_v<FakeIterator<int, int>>);
 static_assert(dpl_unseq::__is_value_storable_v<FakeIterator<std::pair<int, int>, std::pair<int&, int&>>>);
-static_assert(dpl_unseq::__is_value_storable_v<FakeIterator<CopyOnlyNoMove, CopyOnlyNoMove>>);
+static_assert(
+    dpl_unseq::__is_value_storable_v<FakeIterator<TestUtils::CopyOnlyNoMoveCompare, TestUtils::CopyOnlyNoMoveCompare>>);
 // Accepted although the bricks do not compile for them: these iterators do not meet the requirements of a forward
 // iterator, which is not detected here.
-static_assert(std::is_copy_constructible_v<ExplicitCopyCtor>);
-static_assert(dpl_unseq::__is_value_storable_v<ExplicitCopyCtor*>);
-static_assert(dpl_unseq::__is_value_storable_v<ConstCopyOnly*>);
+static_assert(std::is_copy_constructible_v<TestUtils::ExplicitCopyCtorCompare>);
+static_assert(dpl_unseq::__is_value_storable_v<TestUtils::ExplicitCopyCtorCompare*>);
+static_assert(dpl_unseq::__is_value_storable_v<TestUtils::ConstCopyOnlyCompare*>);
 static_assert(dpl_unseq::__is_value_storable_v<FakeIterator<int, OpaqueRef>>);
 
 // Rejected because of the value type: default construction (the first two), copy assignment, copy construction.
 static_assert(!dpl_unseq::__is_value_storable_v<TestUtils::NoDefaultCtorWrapper<int>*>);
-static_assert(!dpl_unseq::__is_value_storable_v<BraceInitOnly*>);
-static_assert(!dpl_unseq::__is_value_storable_v<NoCopyAssign*>);
-static_assert(!dpl_unseq::__is_value_storable_v<MoveOnly*>);
+static_assert(!dpl_unseq::__is_value_storable_v<TestUtils::BraceInitOnlyCompare*>);
+static_assert(!dpl_unseq::__is_value_storable_v<TestUtils::NoCopyAssignCompare*>);
+static_assert(!dpl_unseq::__is_value_storable_v<TestUtils::MoveOnlyCompare*>);
 
 // Rejected because an output iterator reports void as its value type.
 static_assert(!dpl_unseq::__is_value_storable_v<std::back_insert_iterator<std::vector<int>>>);
