+++ test/parallel_api/algorithm/alg.sorting/alg.min.max/minmax_element.pass.cpp
@@ -276,150 +276,10 @@ test_by_type(std::size_t n)
     }
 }
 
-// should provide minimal requirements only
-struct OnlyLessCompare
-{
-    std::int32_t val;
-    OnlyLessCompare() : val(0) {}
-    OnlyLessCompare(std::int32_t val_) : val(val_) {}
-    bool
-    operator<(const OnlyLessCompare& other) const
-    {
-        return val < other.val;
-    }
-};
-
-// Default-constructible through an explicit default constructor only.
-struct ExplicitDefaultCtorCompare
-{
-    std::int32_t val;
-    explicit ExplicitDefaultCtorCompare() : val(0) {}
-    ExplicitDefaultCtorCompare(std::int32_t val_) : val(val_) {}
-    bool
-    operator<(const ExplicitDefaultCtorCompare& other) const
-    {
-        return val < other.val;
-    }
-};
-
-// Default-constructible, but not brace-initializable: empty braces copy-list-initialize the member, which its explicit
-// default constructor rejects.
-struct AggregateOfExplicitDefaultCtorCompare
-{
-    ExplicitDefaultCtorCompare member;
-    bool
-    operator<(const AggregateOfExplicitDefaultCtorCompare& other) const
-    {
-        return member < other.member;
-    }
-};
-
-// Not default-constructible: neither constructor takes zero arguments. Empty braces select the initializer-list one.
-struct BraceInitOnlyCompare
-{
-    std::int32_t val;
-    BraceInitOnlyCompare(std::initializer_list<std::int32_t> init) : val(init.size() == 0 ? 0 : *init.begin()) {}
-    BraceInitOnlyCompare(std::int32_t val_) : val(val_) {}
-    bool
-    operator<(const BraceInitOnlyCompare& other) const
-    {
-        return val < other.val;
-    }
-};
-
-// Copyable, but with deleted move operations.
-struct CopyOnlyNoMoveCompare
-{
-    std::int32_t val;
-    CopyOnlyNoMoveCompare() : val(0) {}
-    CopyOnlyNoMoveCompare(std::int32_t val_) : val(val_) {}
-    CopyOnlyNoMoveCompare(const CopyOnlyNoMoveCompare&) = default;
-    CopyOnlyNoMoveCompare&
-    operator=(const CopyOnlyNoMoveCompare&) = default;
-    CopyOnlyNoMoveCompare(CopyOnlyNoMoveCompare&&) = delete;
-    CopyOnlyNoMoveCompare&
-    operator=(CopyOnlyNoMoveCompare&&) = delete;
-    bool
-    operator<(const CopyOnlyNoMoveCompare& other) const
-    {
-        return val < other.val;
-    }
-};
-
-// The copy assignment returns void instead of VoidAssignCompare&.
-struct VoidAssignCompare
-{
-    std::int32_t val;
-    VoidAssignCompare() : val(0) {}
-    VoidAssignCompare(std::int32_t val_) : val(val_) {}
-    void
-    operator=(const VoidAssignCompare& other)
-    {
-        val = other.val;
-    }
-    bool
-    operator<(const VoidAssignCompare& other) const
-    {
-        return val < other.val;
-    }
-};
-
-// Copyable and assignable from a const lvalue only, so it requires const iterators.
-struct ConstCopyOnlyCompare
-{
-    std::int32_t val;
-    ConstCopyOnlyCompare() : val(0) {}
-    ConstCopyOnlyCompare(std::int32_t val_) : val(val_) {}
-    ConstCopyOnlyCompare(const ConstCopyOnlyCompare&) = default;
-    ConstCopyOnlyCompare(ConstCopyOnlyCompare&) = delete;
-    ConstCopyOnlyCompare&
-    operator=(const ConstCopyOnlyCompare&) = default;
-    ConstCopyOnlyCompare&
-    operator=(ConstCopyOnlyCompare&) = delete;
-    bool
-    operator<(const ConstCopyOnlyCompare& other) const
-    {
-        return val < other.val;
-    }
-};
-
-// A type that is not default-constructible is taken from the test utilities:
-// TestUtils::NoDefaultCtorWrapper<std::int32_t>. It compares through its conversion to the underlying type.
-
-// Not copy-assignable.
-struct NoCopyAssignCompare
-{
-    std::int32_t val;
-    NoCopyAssignCompare() : val(0) {}
-    NoCopyAssignCompare(std::int32_t val_) : val(val_) {}
-    NoCopyAssignCompare(const NoCopyAssignCompare&) = default;
-    NoCopyAssignCompare&
-    operator=(const NoCopyAssignCompare&) = delete;
-    bool
-    operator<(const NoCopyAssignCompare& other) const
-    {
-        return val < other.val;
-    }
-};
-
-// Not copy-constructible.
-struct MoveOnlyCompare
-{
-    std::int32_t val;
-    MoveOnlyCompare() : val(0) {}
-    MoveOnlyCompare(std::int32_t val_) : val(val_) {}
-    MoveOnlyCompare(MoveOnlyCompare&&) = default;
-    MoveOnlyCompare&
-    operator=(MoveOnlyCompare&&) = default;
-    MoveOnlyCompare(const MoveOnlyCompare&) = delete;
-    MoveOnlyCompare&
-    operator=(const MoveOnlyCompare&) = delete;
-    bool
-    operator<(const MoveOnlyCompare& other) const
-    {
-        return val < other.val;
-    }
-};
+// The value types with restricted operations that the test runs the algorithms on - OnlyLessCompare and the rest, plus
+// TestUtils::NoDefaultCtorWrapper<std::int32_t>, which is not default-constructible - are defined in
+// test/support/utils.h, because test/general/implementation_details/value_storable.pass.cpp checks the very same set
+// against the trait that selects the vector code path.
 
 template <typename T, typename Iterator>
 static void
@@ -450,7 +310,7 @@ test_by_type_host_policies(std::size_t n)
         data.emplace_back(std::int32_t(TestUtils::HashBits(i, 30)));
 
     using Iterator = std::conditional_t<UseConstIterators, typename std::vector<T>::const_iterator,
-                                          typename std::vector<T>::iterator>;
+                                        typename std::vector<T>::iterator>;
     check_by_type_host_policies<T>(Iterator(data.begin()), Iterator(data.end()));
 }
